+++ src/neo4j_graphrag/experimental/components/kg_writer.py
@@ -29,10 +29,9 @@ from neo4j_graphrag.experimental.components.types import (
 )
 from neo4j_graphrag.experimental.pipeline.component import Component, DataModel
 from neo4j_graphrag.neo4j_queries import (
-    UPSERT_NODE_QUERY,
-    UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
-    UPSERT_RELATIONSHIP_QUERY,
-    UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
+    upsert_node_query,
+    upsert_relationship_query,
+    db_cleaning_query,
 )
 from neo4j_graphrag.utils.version_utils import (
     get_version,
@@ -117,16 +116,19 @@ class Neo4jWriter(KGWriter):
         driver: neo4j.Driver,
         neo4j_database: Optional[str] = None,
         batch_size: int = 1000,
+        clean_db: bool = True,
     ):
         self.driver = driver_config.override_user_agent(driver)
         self.neo4j_database = neo4j_database
         self.batch_size = batch_size
+        self._clean_db = clean_db
         version_tuple, _, _ = get_version(self.driver, self.neo4j_database)
         self.is_version_5_23_or_above = is_version_5_23_or_above(version_tuple)
 
     def _db_setup(self) -> None:
-        # not used for now
-        pass
+        self.driver.execute_query("""
+        CREATE INDEX __entity__tmp_internal_id IF NOT EXISTS FOR (n:__KGBuilder__) ON (n.__tmp_internal_id)
+        """)
 
     @staticmethod
     def _nodes_to_rows(
@@ -144,61 +146,52 @@ class Neo4jWriter(KGWriter):
 
     def _upsert_nodes(
         self, nodes: list[Neo4jNode], lexical_graph_config: LexicalGraphConfig
-    ) -> dict[str, str]:
-        """Upserts a single node into the Neo4j database."
+    ) -> None:
+        """Upserts a batch of nodes into the Neo4j database.
 
         Args:
             nodes (list[Neo4jNode]): The nodes batch to upsert into the database.
         """
         parameters = {"rows": self._nodes_to_rows(nodes, lexical_graph_config)}
-        query = (
-            UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE
-            if self.is_version_5_23_or_above
-            else UPSERT_NODE_QUERY
-        )
-        records, _, _ = self.driver.execute_query(
-            query,
-            parameters_=parameters,
-            database_=self.neo4j_database,
-        )
-        return {r["_internal_id"]: r["element_id"] for r in records}
-
-    @staticmethod
-    def _relationships_to_rows(
-        relationships: list[Neo4jRelationship], node_id_mapping: dict[str, str]
-    ) -> list[dict[str, Any]]:
-        return [
-            {
-                **relationship.model_dump(),
-                "start_node_element_id": node_id_mapping.get(
-                    relationship.start_node_id, ""
-                ),
-                "end_node_element_id": node_id_mapping.get(
-                    relationship.end_node_id, ""
-                ),
-            }
-            for relationship in relationships
-        ]
-
-    def _upsert_relationships(
-        self, rels: list[Neo4jRelationship], node_id_mapping: dict[str, str]
-    ) -> None:
-        """Upserts a single relationship into the Neo4j database.
-
-        Args:
-            rels (list[Neo4jRelationship]): The relationships batch to upsert into the database.
-        """
-        parameters = {"rows": self._relationships_to_rows(rels, node_id_mapping)}
-        query = (
-            UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE
-            if self.is_version_5_23_or_above
-            else UPSERT_RELATIONSHIP_QUERY
+        query = upsert_node_query(
+            support_variable_scope_clause=self.is_version_5_23_or_above
         )
         self.driver.execute_query(
             query,
             parameters_=parameters,
             database_=self.neo4j_database,
         )
+        return None
+
+    @staticmethod
+    def _relationships_to_rows(
+        relationships: list[Neo4jRelationship],
+    ) -> list[dict[str, Any]]:
+        return [relationship.model_dump() for relationship in relationships]
+
+    def _upsert_relationships(self, rels: list[Neo4jRelationship]) -> None:
+        """Upserts a batch of relationships into the Neo4j database.
+
+        Args:
+            rels (list[Neo4jRelationship]): The relationships batch to upsert into the database.
+        """
+        parameters = {"rows": self._relationships_to_rows(rels)}
+        query = upsert_relationship_query(
+            support_variable_scope_clause=self.is_version_5_23_or_above
+        )
+        self.driver.execute_query(
+            query,
+            parameters_=parameters,
+            database_=self.neo4j_database,
+        )
+
+    def _db_cleaning(self) -> None:
+        query = db_cleaning_query(
+            support_variable_scope_clause=self.is_version_5_23_or_above,
+            batch_size=self.batch_size,
+        )
+        with self.driver.session() as session:
+            session.run(query)
 
     @validate_call
     async def run(
@@ -215,14 +208,14 @@ class Neo4jWriter(KGWriter):
         try:
             self._db_setup()
 
-            node_id_mapping = {}
-
             for batch in batched(graph.nodes, self.batch_size):
-                batch_mapping = self._upsert_nodes(batch, lexical_graph_config)
-                node_id_mapping.update(batch_mapping)
+                self._upsert_nodes(batch, lexical_graph_config)
 
             for batch in batched(graph.relationships, self.batch_size):
-                self._upsert_relationships(batch, node_id_mapping)
+                self._upsert_relationships(batch)
+
+            if self._clean_db:
+                self._db_cleaning()
 
             return KGWriterModel(
                 status="SUCCESS",